projects/0757ea4c-5bb8-4630-82c5-5d15f380c447/archive/20230815_184532/workspace/requirements.txt
flask==1.1.2
requests==2.25.1
python-dotenv==0.15.0
pytest==6.2.2
